硬件不支持报警

+++ ametal_am824_core_1.02/projects_keil5/applications/template/user_code/demo_temp.c
@@ -32,7 +32,7 @@ void demo_lm75_temp_entry (void)
     
     am_temp_info_get(0, &info);
     am_temp_read(0, &value);
-    AM_DBG_INFO("current tempture is %.2f\n", value / info.multiple);
+    AM_DBG_INFO("current tempture is %.2f\n", value / (float)info.multiple);
 }
  
 /* end of file */
